# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.transactionSummaryData:
    additionalProperties: true
    type: object
  requests.CreateBankAccountRequest:
    properties:
      account_name:
        type: string
      account_type:
        type: string
      balance:
        type: number
      category:
        type: string
      currency:
        type: string
      user_id:
        type: integer
    type: object
...
  responses.AccountResponse:
    properties:
      account_type:
        type: string
      balance:
        type: number
      category:
        type: string
      currency:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  responses.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  responses.TransactionResponse:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      category:
        type: string
      currency:
        type: string
      date:
        type: integer
      description:
        type: string
      id:
        type: integer
      payee:
        type: string
      reference:
        type: string
      transaction_status:
        type: string
      transaction_type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Haven Finance API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Haven Finance API
  version: "1.0"
paths:
  /accounts:
    get:
      description: Retrieve all accounts
      parameters:
      - description: Group by account type
        enum:
        - true
        - false
        in: query
        name: group_by_account_type
        type: boolean
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.AccountResponse'
            type: array
      security:
      - AuthToken: []
      summary: Get all accounts
      tags:
      - accounts
  /accounts/{id}:
    get:
      description: Retrieve an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.AccountResponse'
      security:
      - AuthToken: []
      summary: Get an account
      tags:
      - accounts
  /accounts/{id}/transactions:
    get:
      description: Retrieve an account's transactions
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.TransactionResponse'
            type: array
      security:
      - AuthToken: []
      summary: Get an account's transactions
      tags:
      - accounts
  /accounts/{id}/transactions/recent:
    get:
      consumes:
      - application/json
      description: Get 5 recent transactions for an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.TransactionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - AuthToken: []
      summary: Get 5 recent transactions for an account
      tags:
      - accounts
  /accounts/{id}/transactions/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload account transactions
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Transactions File
        in: formData
        name: file
        required: true
        type: file
      - description: Transaction Schema
        in: formData
        name: transaction_schema
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.TransactionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - AuthToken: []
      summary: Upload account transactions
      tags:
      - accounts
  /accounts/create:
    post:
      consumes:
      - application/json
      description: Create an account
      parameters:
      - description: Create Account Request
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/requests.CreateBankAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - AuthToken: []
      summary: Create an account
      tags:
      - accounts
...
  /data/{account_id}/transactions/histogram:
    get:
      description: Get transaction histogram data
      operationId: get-transaction-histogram-data
      parameters:
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: integer
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - AuthToken: []
      summary: Get transaction histogram data
      tags:
      - data
  /data/{account_id}/transactions/summary:
    get:
      description: Get transactions summary data
      parameters:
      - description: Filter
        enum:
        - transaction_category
        - transaction_type
        - month
        - year
        - week
        in: query
        name: filter
        type: string
      - description: Interval
        enum:
        - month
        - year
        - week
        in: query
        name: interval
        type: string
      - description: Group by
        in: query
        name: group_by
        type: string
      - description: Sort by
        in: query
        name: sort_by
        type: string
      - description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: integer
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.transactionSummaryData'
            type: array
      security:
      - AuthToken: []
      summary: Get transactions summary data
      tags:
      - data
...
swagger: "2.0"